Changed the datatype returned by Servo.getPosition() to be byte. Created new array in class enviroment scaner which will be storing data from readings.

+++ tests/pilot/src/Servo.java
@@ -14,9 +14,10 @@ public class Servo extends Thread {
 			if (!_suppressed) servo.rotateTo(Const.SERVO_MAX_ANGLE);
 		}
 	}
-
-	public float getPosition() {
-		return servo.getTachoCount();
+	
+	// It is now returning byte for sake of saving memory
+	public byte getPosition() {
+		return (byte) servo.getTachoCount();
 	}
 	
 	public void turnOff() {
